refactor(sentry): to generate prop-types from TS
