+++ packages/sentry/src/index.ts
@@ -1,6 +1,10 @@
 export { default as version } from './version';
 export * from './sentry';
-export { default as SentryUserTracker } from './sentry-user-tracker';
+export {
+  default as SentryUserTracker,
+  updateUser,
+} from './sentry-user-tracker';
 export {
   default as SentryUserLogoutTracker,
+  stopTrackingUser,
 } from './sentry-user-logout-tracker';
