+++ packages/sentry/src/sentry-user-logout-tracker/index.ts
@@ -1,1 +1,1 @@
-export { default } from './sentry-user-logout-tracker';
+export { default, stopTrackingUser } from './sentry-user-logout-tracker';
