+++ packages/sentry/src/sentry-user-tracker/index.ts
@@ -1,1 +1,1 @@
-export { default } from './sentry-user-tracker';
+export { default, updateUser } from './sentry-user-tracker';
