+++ packages/sentry/src/sentry-user-tracker/sentry-user-tracker.tsx
@@ -1,22 +1,35 @@
 import React from 'react';
-import * as PropTypes from 'prop-types';
-import { InferPropTypes } from '../type-utils';
-import * as sentry from '../sentry';
+import * as Sentry from '@sentry/browser';
 
-const userTrackerPropTypes = {
-  user: PropTypes.shape(sentry.userPropTypes),
+type SentryUser = {
+  id: string;
+  email: string;
+};
+type Props = {
+  user?: SentryUser;
 };
 
-type UserTrackerProps = InferPropTypes<typeof userTrackerPropTypes>;
+export const updateUser = (user: SentryUser) => {
+  if (window.app.trackingSentry) {
+    // to avoid sending personal data to sentry we anonymize the email address
+    // by only sending the domain part or the email
+    const emailTld = user.email.split('@')[1];
+    Sentry.configureScope(scope => {
+      scope.setUser({
+        email: `xxx@${emailTld}`,
+        id: user.id,
+      });
+    });
+  }
+};
 
 /**
  * This component will let sentry know if any information about the user has
  * changed.
  */
 
-class SentryUserTracker extends React.PureComponent<UserTrackerProps> {
+class SentryUserTracker extends React.PureComponent<Props> {
   static displayName = 'SentryUserTracker';
-  static propTypes = userTrackerPropTypes;
   componentDidMount() {
     // since the user and project could have been loaded from the apollo cache
     // they could be preset already when mounting
@@ -27,7 +40,7 @@ class SentryUserTracker extends React.PureComponent<UserTrackerProps> {
   }
   syncUser = () => {
     if (this.props.user) {
-      sentry.updateUser(this.props.user);
+      updateUser(this.props.user);
     }
   };
   render() {
